fmt_scan.h: Base header on implementation.
Normally done as part of extracting the new header file from a larger
one, but this file already existed so it had to be moved as one commit
and altered in the next, otherwise git would break its history.

Move structs comp and format, and many prototypes, to existing
fmt_compile.h because the functions are in fmt_compile.c.  Change
cpnumber() and cptrimmed() to be static as they've no callers outside
fmt_scan.c.