patch: Maildir support for "scan -file"
I have extended the "scan -file" command line so that the "file" can
also be a Maildir directory.

This feature complements the Maildir support added to the inc command
in 1.4.  Just as you may want to inc your mail from Maildir, you might
want to see what awaits you with a scan first.

Signed-off-by: David Levine <levinedl@acm.org>