Fixed memory corruption in post(1) and inc(1) when using XOAUTH2.
An internal table for the oauth entries was initially sized at 4,
and then reallocated if the user had more than 4.  However, the
reallocation didn't account for the size of the allocated
structure.

Also, increased the initial size to 16, trading off slightly
increased memory usage for performance.

Fix to commit 29ff4f879.