uip/mhparse.c: fix SIGSEGV when ‘<#’ content-type has missing slash.
If the content-type line being parsed by get_ctinfo() was a ‘#<’
directive and didn't have a slash after the content-type's type then no
subtype had to be supplied and ci_subtype was left unaltered.

user_content() initialises ci_subtype to zero bytes which is typically
a NULL pointer before calling get_ctinfo() and, later, build_headers()
calls strlen(ci_subtype) without checking if it may be NULL first,
causing the SIGSEGV.

get_ctinfo() is given a string to parse with the leading directive
stripped off, e.g. ‘begin’ instead of ‘#begin’.  It complained
of a missing type on ‘’ and a missing subtype on ‘foo/’ because
the slash indicated a subtype should follow, but ‘foo’ was returned
as the type with a NULL subtype to allow for ‘begin’, etc.  Instead,
explicitly test for the three directives which don't require a slash,
i.e. ‘forw’, ‘begin’, and ‘end’, and otherwise report a
missing-slash error.

This area has changed before.  If the slash was missing then mh-6.8.5's
uip/mhn.c 1.1 complained, assuming no ‘magic’ mode.  In 1.2, it
changed to setting subtype to an empty string if not in magic mode and
always skipping attempting to parse any subtype; the old error message
had been preprocessed out.  The final change was in 2.0 where the error
message was removed and the existing logic tidied.

Add some initial tests of ‘#<’ parsing by mhbuild, including the
slash being missing.  Much more is required.