sbr/fmt_rfc2047.c: avoid buffer overflow when iconv(3) not used.
If iconv(3) is available then when the available space in the caller's
output buffer reached zero, measured by ‘dstlen’, decode_rfc2047()'s
ADDCHR2() macro would jump to the ‘iconvbuffull’ label which would do
nothing if iconv() wasn't needed for the character set allow the loop to
continue past the end of the buffer.

The fix is to spot iconv isn't being used and jump to the normal
‘buffull’ label.

This was causing mhl(1) to SIGSEGV on an email with lots of Subject
fields where the total field values needed more than the 8 KiB buffer to
display them all as one Subject field.