all *.[ch]: turn ‘debugsw’ from int to bool to match its values.
Some library routines expect a global ‘debugsw’ to exist thus many
programs have to define it.  They all agreed on using int but it was
only ever tested for truthness and its value ignored.  Switch it to
being a bool for clarity.  This removes the possibility of having a
higher level of debug but if that's needed then the whole area needs
reworking anyway, as an existing comment or two says.

The parameter to a few list_...() functions changes too as ‘debugsw’ is
passed to them and then used by them as a Boolean.