sbr/path.c: reworked etcpath() of paths that start with ~.
The initial motivation was to clean up a couple of warnings with
-Wstringop-overflow from gcc 11 with optimization enabled.  For
an input of ~user/file when that user does not exist, the old code
would restore the first '/' after having removed it earlier, to
isolate ~user.  Changed the code to not remove it if that user
doesn't exist.

Added test that exercises that code to test/mhical/test-mhical.