uip/mkstemp.c: use memcpy(3) instead of strncpy(3) to hush compiler.
strncpy() was being used to copy the source string apart from its
terminating NUL, and sometimes the source string could be empty so the
destination wouldn't be written at all; gcc noticed.

The length to copy is known, there is no risk of stopping early on an
unexpected NUL, so use memcpy() to clarify intent.

Also, avoid copying the NUL of prefix[].