sbr/netsec.c: hush compiler by replacing fprintf() with fwrite().
A printf(3) precision-specifier of ‘*’ takes an int but an unsigned was
being passed.  Given the intent of ‘"%.*s\n", outlen, snoopoutbuf’ was
to write outlen bytes of snoopoutbuf and a linefeed, replace it with an
fwrite() and putc().  Examining the code shows no NUL is expected in
snoopoutbuf within its first outlen bytes so using ‘%s’ is a bit
misleading in isolation to the reader and makes more work for stdio.