uip/mhfixmsg.c: hush compiler by enlarging snprintf() destination.
Alter the destination array's definition to use sizeof on a string
constant which shows the maximum which could be produced.  Don't use
stdint.h's SIZE_MAX in the string as that comes with extra noise like
a trailing ‘UL’ and parenthesis.

A size_t is being printed so use ‘%zu’ rather than cast to unsigned.