sbr/utils.c: add abortcpy() which prefers to abort than to overflow.
This is the partner to trunccpy().  If truncation is valid, e.g. passing
on the first part of a long string in an error message, then trunccpy()
should be used instead of strncpy(3).  Else abortcpy() should be used
as it will abort(3) rather than overflow the destination or leave the
destination not terminated.

(strncpy() is rarely wanted as its NUL-padding is only useful to pad
fixed-width fields, e.g. an old 14-byte filename which sits alongside
the 16-bit inode number in a directory entry.)

The creation of abortcpy() and its environment-variable workaround was
discussed on nmh-workers around 2016-10-29, in particular message-ID
<20161029214059.1046247BE8@pb-smtp2.pobox.com> at 21:40:58 UTC.