sbr/path.c: replace will-never-truncate TRUNCCPY() with memcpy().
No truncation can occur because the source's length is obtained with
strlen() before the TRUNCCPY() and explicitly checked against the
destination's size in order to give an explicit stderr warning.  As the
source length is known and it will fit, just memcpy the bytes rather
than use TRUNCCPY() which will duplicate finding the string's length.