uip/imaptest.c: replace mh_xmalloc() and trunccpy() with xmemtostr().
parse_capability() takes a pointer to a run of char and their length; they
may not be NUL-terminated.  The code was and still does copy them into
newly malloc'd memory and appends a NUL.  The copy was using trunccpy()
knowing it would copy all the needed source bytes before stopping as no
source NUL was found and putting its own NUL at the end, but that's a bit
obscure and duplicates effort as the number of bytes to copy is known.
Instead, use the new xmemtostr() to replace both calls.  It also has
the advantage of spotting an overflow of the length, say when pointer
arithmetic produces -1 AKA SIZE_MAX without which the bytes would be
copied to mh_xmalloc(0).