uip/flist.c: replace literal tab in string with three spaces.
A string of spaces was being printed in lieu of ‘ (private)’ to maintain
columnar output.  This was changed, probably by a faulty editor
configuration, to replace the initial spaces with a tab character.
The source code's layout didn't change, but the tab isn't sufficient to
keep the columns aligned in flist's output.

Fixes e491542d.